requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0
pyodbc>=4.0.39
pymysql>=1.1.0

# Authentication & Security
passlib[bcrypt]>=1.7.4
bcrypt<4
python-jose[cryptography]>=3.3.0

# Cloud Storage & APIs
azure-storage-blob>=12.19.0
openai>=1.30.0
google-generativeai>=0.3.2
google-api-python-client

# Utilities
python-dotenv>=1.0.1
requests>=2.31.0
email-validator>=2.1.0.post1

# Image & PDF Processing
Pillow>=10.4.0
reportlab>=4.2.2

# Machine Learning & Data Science
torch
torchvision
scikit-learn>=1.3.0
numpy

# Visualization
matplotlib>=3.8.0
seaborn

# Data Collection
pydrive2
beautifulsoup4

# Data Models
pydantic
